Change in-memory storage for the DataTracker.
Instead of caching everything in a series of maps, instead cache
things in a series of slices that we keep in order.  This will make it
easier to deal with the DHCP code in the future, as well as making it
easier for the API to deal with pagination at some point in the future.