Add functions for finding and creating Leases.
This consists of two functions:

FindLease should be used in response to a DHCPREQUEST.  It
encapsulates the logic needed to find a lease to renew, or to signal
that the lease cannot be renewed.

FindOrCreateLease should be used in response to a DHCPDISCOVER.  It
encapsulates the logic needed to find an extant lease for an
interface, create a new one if one was not found, or signal that the
DHCP system should not respond due to either not finding a lease or
being unable to create one for the system.