Add handling for handing out existing leases in findViaSubnet
We were missing the case where we handed out a lease, then got a
DHCPDISCOVER shortly thereafter from the same node -- we would always
hand out a different lease in that case.

Add handling that detects when an incoming request for a lease that is
not otherwise satisfied by a reservation can be satisfied by an existing
non-expired lease in the appropriate subnet. If it can, we will use that
lease without referring to any of that subnet's Pickers.

Tests added to cover this case.