Make main responsible for extracting assets.
The backend had been handling it mostly because nothing else was present
when the asset handling code was added.  Now that we have an actual
non-stub main package, make it responsible for expanding any assets we
need at runtime.

This also does a bit of refactoring to consolidate when and how binary
assets are downloaded -- instead of it being a seperate step, it is part
of go generate now.