Rework DHCP server to be more flexible
You can now tell the DHCP server to only listen to requests on specific
interfaces.  This is useful when you don't want to inadvertently stomp
all over someone else's DHCP server on a nic you shoudn't be paying
attention to.

The DHCP server also no longer ignores secondary addresses on network
interfaces -- if you have a nic with two IP addresses in different
subnets bound to it, the DHCP packet handling code will use all of them
to determine subnet eligibility and what address to use as the server
address for that particular DHCP interaction.

Also, NextServer is always set now, allowing PXE boots to actually
work.  If neither the subnet nor the reservation for a lease specify a
NextServer, it will default to an IP address we are listening on that
best matches the lease's address.