Split mains into the cmds directory. Allow cli and server to run as their own packages. This will allow for unit tests.
Update build and test tools for it.

Also generate darwin and linux clients.  Place them in the
embedded assets.  This is for now.  Probabyl undo this, but it
allows macs to work.  build.sh process rscli in top dir for
build platform along with rocket-skates.

Update environments to use the correct rscli and fix a rebar
cli issue in discovery.