Start reworking RenderData to use the address cache.
This updates the DataTracker to generate FileURL and ApiURL
dynamically, and wires everything else up to use the methods instead
of static struct fields.

This also changes how we reference the URL machines should be
installing from, so an update of all templates and bootenvs is
required.

Functionally, RenderData still winds up statically rendering files to
the disk, and itt does not yet use the address cache.  Using that will
involve using a virtual filesystem that renders templates on demand
instead of prerendering them to the filesystem.  Once that is in
place, we can wire RenderData to use the address cache.