Add Token support
This adds a user/#name/token endpoint that generates a token for access.
The rest of the endpoints validate the token if provided in the request
header, DRP-AUTH-TOKEN.  Individual endpoints will use the
encoded structure to restrict actions based upon the internal values.
The token also has an expiration value as well.