Make template rendering avoid inadvertent data leaks
This refactors how renderData works to avoid inadvertent leaks between
the environment where the bootenvs, machines, and templates are
created and the time they are rendered.  Now, render time explicitly
reloads all pertinent data, which ensures that it always works with
the most current versions of the objects in question.

All that is left to do at object change time is register and
deregister paths and their associated callbacks in the overlay
filesystem.

Also added appropriate hooks to ensure that all appropriate paths are
rerendered when a bootenv changes