Add Unique to indexes, enforce it, and let Frontend use it.
This adds a flag to index.Maker that indicates whether values in an
index should be unique.  That flag is enforced by a BeforeSave hook on
each item in the DataTracker.  FetchOne has been updated to allow
fetching items on a unique index if the key is passed in the format of
index:unique_value