Start adding Tasks into dr-provision
This adds the Tasks model to the backend, and wires it in all the
places it needs to be w.r.t template expansion.

Please note that this is a work in progress -- Tasks are wired in, but
they do not do anything and there are further support changes to be
made for logging, task tracking, etc.