This not so small commit adds the Stage Object.
The Stage Object (and its cli, frontend, and backend components)
handle the tasks and profiles from the bootenvs.

The BootEnv shrinks to be just a boot environment for a machine.

The Stage object now contains the tasks to execute and the profiles
to temporarily configure machines for those tasks.

The machine object is updated to optionally have a stage.  If stage
is not present, the system will continue to work like today.

The profile object loses its task list.  This was an idea looking
for stages.  Profile's with tasks were basically stages in a bad way.
The whole task system gets simpler with the stages system.

The stages render machine templates only.

The machine object has also been updated to allow for the
returning of aggregated params.  The cli and API take a
new flag aggregate=true to return all current parameters for
the machine (from profiles, stage, and machine).