Add content feature flags, and use that to add new exit states for tasks.
All content layers and items in DRP can now specify a meta item called
"feature-flags", which must contain s comma-seperated list of tags
that will be interpreted as feature flags.  The first use for these
feature flags is to allow content laters and individual tasks to
specify which exit code scheme the task runner should use for
interpreting task exit codes.

The original scheme (signalled by the original-exit-codes feature
flag, or by the lack of a sane-exit-codes feature flag) retains the
meaning of the original exit codes:

  0: success
  1: reboot
  2: incomplete
  3: reboot+incomplete
  4 and up: fail

If an individual task does not have a meta elememt called
"feature-flags", then one will be created for it and the
"original-exit-codes" feature flag will be added to it.

The new scheme (indicated by the "sane-exit-codes") feature flag uses
bits 5, 6, and 7 of the exit status to indicate to the task runner the
following special actions to take:

To set that the job is incomplete, exit $(($? | 128)).  This sets bit
7 of the exit code.

To set that the runner should reboot the system, exit $(($? | 64)).
This sets bit 6 of the exit code.

To set that the runner should shut the system down, exit $(($? |
32)). This sets bit 5 of the exit code.

If the reboot and the shutdown bits are both set, the system will
reboot.

If any of the lower bits are set, we will consider the task to have
failed, otherwise we will conside rthe task to have succeeded.

In particular, exit 1 is now just a standard fail exit code, not a
success with reboot code.