This patch adds three secret keys to the JWT auth token. User, Grantor, and Machine secrets.
These secrets are include in the token when sealed.
When validated, the secrets are checked to make sure
that the value hasn't changed. If secrets don't match,
the auth is not granted.  This allows for revocation.

The GenerateToken and new GenerateInfiniteToken commands
use the pref 'systemGrantorSecert' and the machine's
specific secret included in the token.

The Generate user token API method uses the secret on
the granting user's object and the authorized user's object
to populate their secrets.

The new secret fields are generated for existing objects
as well.