Populate the address cache when handling API requests.
This also changes the default static IP from 192.168.124.11 to ``,
and adds a fallback to try and figure out a valid static IP by
figuring out which interface the default route goes through and using it.

dr-provision will log whenever it has to fall back to using --static-ip or
the IP of the interface with the default route instead of the address cache.