Clean up when and how we set DHCP pxe boot options
Always pass a PXE menu with one entry and a 0 second timeout.  Some UEFI
bioses need this to netboot properly.

Detect whenever we are rendering bootparams for an ipxe template and
inject an initrd= parameter.  The ipxe.efi loader needs this, and it
does not appear to harm the legacy BIOS loader.  It is probably a bug
in ipxe.

Whenever we return a DHCP bootfile, also set that bootfile in the
FNAME field in the bootp packet.

Set the TFTP server name field explicitly instead of relying on
fallback to the DHCP server.