Fix up DHCP handler to more closely adhere to the DHCP spec.
RFC2132 implies that we should prefer using the DHCP packet sname and
file fields to supply the TFTP server name and the bootfile instead of
leaving those fields blank and using options 66 and 67.  Most clients
don't care and will use whichever is available, but the UEFI client
baked in to the Dell T320 only seems to work reliably when using the
server name and file fields -- leaving them blank and using options 66
and 67 (or just 67 and the siaddr field) result in the client trying
to fetch the boot file with a bunch of extra junk appended to the
name, or just failing to figure out the address of the tftp server.

Since this is probably a failure inherent in all tianocore defived
UEFO firmware from that era and the rest of the hardware (along with
current qemu running seabios and tianocore) I have access to works
fine with these changes in legacy BIOS mode, this should be a safe
change to make.

This also reverts back to the previous proxy DHCP behaviour of only
adding PXE specific options when we should respond to DHCP discovery
packets as if we were a proxy DHCP server, instead of including the
PXE speecific DHCP options whenever the DHCP request indicated it was
coming from a PXE client.