This addresses a deadlock in the event system. The unit tests would hang but still print messages in travis until it times out after an hour.
The problem is that we test the event registration
system by registering for an event and immediately
sending it.  The problem is that the register send
on the websocket is not blocking and can return
before the registration is complete.  For our
unit test, the sent event is handled before
the registration is complete causing the event
to not be sent on the websocket.

To fix this, the register and deregister calls
generate events.  The event stream register
and deregister calls don't return until the
event from the registration call returns.

This allows for the system not "miss" events during
registration.