Fix busted machine rendering
In the distant past, I foixed everything that can render templates to
only do so after all validation checks had passed and the object was
safely saved onto disk.  Except, for some reason that probably seemed
good at the time, for machines.

This PR corrects that oversight and also adds an addiional check to
verify that machines do not wind up with duplicate IP addresses