In some cases, the machine doesn't advance it task count even though it is done.
An example of this is when the terraform ready stage
as its last task shutdowns the machine.  The runner correctly
finishes the task and marks it as finished, but the runner
never comes back around to see if there are more tasks to run.
It knows the it is done with this task list because it marks it
as finished instead of incomplete.  Everything works fine, but
the user must force a stage change to leave this stage (the
shutdown comes before stage changes).

This change advances the currenttask to indicate a completed task
list if the job is marked as finished and the task is the last one.

This doesn't bother the runner or the frontend as they already handle
these cases.