Rework package-repositories handling to prefer local install repos.
Except on debianoids, becasuse recent Ubuntu systems require an
InRelease file these days instead of Release and Release.gpg on the
installed systems, but the install disks are not packaged with the new
hotness, so there is no non-heuristic way to tell.  We will also need
kernel lookaside in package-repositories for completely ISO-free OS
installation purposes, because debian and ubuntu are still weird about
ISO repo layout vs. repos on the internets.

Centos, as usual, continues to Just Work(tm).

Presumably because detached Release and Release.gpg