Use net.Pipe to implement the logger for the Agent
Since it is purely in-memory, synchronous, and implements ReadTimeout,
it can be used with a minimal amount of buffering to eliminate the
special case where os.File does not implement ReadTimeout.

This also adds a CLI unit test to make sure log coalescing is
operating properly in the face of lots of volume and pauses.