Add warm standby HA support.
This allows a consul-based DRP to run multiple instances
of DRP on the same subnet and move a VIP between them
based upon the leader election in consul.  If a DRP process
exits or dies, the leader election continues to choose
another DRP, moves the VIP, and continues on the path. If
the leader DRP exits cleanly, it will release the leadership
role and new one will be elected.

This does NOT sync the tftpboot, plugins, or saas-content
directories.  This is the missing piece for an MVP HA
DRP system.